pyproject.toml
[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[project]
# PyPI/项目显示名（可改）。包代码在 src/biai_server 下。
#
# 说明：
# - uv_build 默认会用 [project].name（规范化后）推断 src/ 下的包目录
# - 为了只保留一个源码包 `biai_server`（不需要额外薄包装包），这里将发行名改为
#   会规范化为 `biai_server` 的名字：biai-server
name = "biai-server"
version = "0.1.0"
description = "BIAI 知识库 MCP 服务器（表结构/业务需求向量检索）"
# 重要：
# 你当前本地 `uv sync` 选用了 CPython 3.14（见终端输出），这会触发部分依赖在
# python>=3.13 的条件依赖（例如 langchain-chroma==1.0.0 需要 numpy>=2.1.0），
# 与我们从 requirements 全量迁移来的 numpy==1.26.4 发生冲突。
# 为了“全量锁版本复刻原环境”并与 Smithery（Python 3.12+）保持一致，这里将
# Python 版本约束到 3.12.x。
requires-python = ">=3.12,<3.13"

# 说明：
# - 这里已按你的要求，将原 requirements.txt **全量锁版本** 迁移到 pyproject.toml
# - 这会包含大量“传递依赖”，体积较大，但可最大程度复现原环境
dependencies = [
  "aiohappyeyeballs==2.6.1",
  "aiohttp==3.13.2",
  "aiosignal==1.4.0",
  "annotated-doc==0.0.4",
  "annotated-types==0.7.0",
  "anthropic==0.75.0",
  "anyio==4.12.0",
  "art==6.5",
  "asgiref==3.11.0",
  "attrs==25.4.0",
  "Authlib==1.6.6",
  "backoff==2.2.1",
  "backports.tarfile==1.2.0",
  "bcrypt==5.0.0",
  "beartype==0.22.9",
  "beautifulsoup4==4.14.3",
  "bracex==2.6",
  "bs4==0.0.2",
  "build==1.3.0",
  "cachetools==6.2.2",
  "certifi==2025.11.12",
  "cffi==2.0.0",
  "charset-normalizer==3.4.4",
  "chroma-hnswlib==0.7.3",
  "chromadb==1.3.6",
  "click==8.3.1",
  "cloudpickle==3.1.2",
  "coloredlogs==15.0.1",
  "cryptography==46.0.3",
  "cyclopts==4.4.0",
  "dataclasses-json==0.6.7",
  "deepagents==0.3.0",
  "diskcache==5.6.3",
  "distro==1.9.0",
  "dnspython==2.8.0",
  "docstring_parser==0.17.0",
  "docutils==0.22.3",
  "durationpy==0.10",
  "email-validator==2.3.0",
  "exceptiongroup==1.3.1",
  "fakeredis==2.33.0",
  "fastapi==0.124.2",
  "fastmcp==2.14.1",
  "filelock==3.20.0",
  "flatbuffers==25.9.23",
  "frozenlist==1.8.0",
  "fsspec==2025.12.0",
  "google-api-core==2.28.1",
  "google-api-python-client==2.187.0",
  "google-auth==2.41.1",
  "google-auth-httplib2==0.3.0",
  "google-auth-oauthlib==1.2.3",
  "googleapis-common-protos==1.72.0",
  "grpcio==1.76.0",
  "h11==0.16.0",
  "hf-xet==1.2.0",
  "httpcore==1.0.9",
  "httplib2==0.31.0",
  "httptools==0.7.1",
  "httpx==0.28.1",
  "httpx-sse==0.4.3",
  "huggingface_hub==1.2.2",
  "humanfriendly==10.0",
  "idna==3.11",
  "importlib_metadata==8.7.0",
  "importlib_resources==6.5.2",
  "jaraco.classes==3.4.0",
  "jaraco.context==6.0.1",
  "jaraco.functools==4.3.0",
  "jiter==0.12.0",
  "jsonpatch==1.33",
  "jsonpointer==3.0.0",
  "jsonschema==4.25.1",
  "jsonschema-path==0.3.4",
  "jsonschema-specifications==2025.9.1",
  "keyring==25.7.0",
  "kubernetes==34.1.0",
  "langchain==1.1.3",
  "langchain-anthropic==1.3.0",
  "langchain-chroma==1.0.0",
  "langchain-classic==1.0.0",
  "langchain-community==0.4.1",
  "langchain-core==1.2.0",
  "langchain-openai==1.1.1",
  "langchain-text-splitters==1.0.0",
  "langgraph==1.0.4",
  "langgraph-checkpoint==3.0.1",
  "langgraph-prebuilt==1.0.5",
  "langgraph-sdk==0.2.15",
  "langsmith==0.4.59",
  "lupa==2.6",
  "markdown-it-py==4.0.0",
  "marshmallow==3.26.1",
  "mcp==1.24.0",
  "mdurl==0.1.2",
  "mmh3==5.2.0",
  "more-itertools==10.8.0",
  "mpmath==1.3.0",
  "multidict==6.7.0",
  "mypy_extensions==1.1.0",
  "numpy==1.26.4",
  "oauthlib==3.3.1",
  "onnxruntime==1.23.2",
  "openai==2.9.0",
  "openapi-pydantic==0.5.1",
  "opentelemetry-api==1.39.1",
  "opentelemetry-exporter-otlp-proto-common==1.39.0",
  "opentelemetry-exporter-otlp-proto-grpc==1.39.0",
  "opentelemetry-exporter-prometheus==0.60b1",
  "opentelemetry-instrumentation==0.60b1",
  # 与 opentelemetry-semantic-conventions==0.60b1 对齐，避免 b0 强制依赖 ==0.60b0 导致不可解
  "opentelemetry-instrumentation-asgi==0.60b1",
  "opentelemetry-instrumentation-fastapi==0.60b1",
  "opentelemetry-proto==1.39.0",
  "opentelemetry-sdk==1.39.1",
  "opentelemetry-semantic-conventions==0.60b1",
  "opentelemetry-util-http==0.60b1",
  "orjson==3.11.5",
  "ormsgpack==1.12.0",
  "overrides==7.7.0",
  "packaging==23.2",
  "pathable==0.4.4",
  "pathvalidate==3.3.1",
  "platformdirs==4.5.1",
  "posthog==5.4.0",
  "prometheus_client==0.23.1",
  "propcache==0.4.1",
  "proto-plus==1.27.0",
  "protobuf==6.33.2",
  "pulsar-client==3.8.0",
  "py-key-value-aio==0.3.0",
  "py-key-value-shared==0.3.0",
  "pyasn1==0.6.1",
  "pyasn1_modules==0.4.2",
  "pybase64==1.4.3",
  "pycparser==2.23",
  "pydantic==2.12.5",
  "pydantic-settings==2.12.0",
  "pydantic_core==2.41.5",
  "pydocket==0.15.6",
  "Pygments==2.19.2",
  "PyJWT==2.10.1",
  "PyMySQL==1.1.2",
  "pyparsing==3.3.1",
  "pyperclip==1.11.0",
  "PyPika==0.48.9",
  "pyproject_hooks==1.2.0",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.2.1",
  "python-json-logger==4.0.0",
  "python-multipart==0.0.20",
  "PyYAML==6.0.3",
  "redis==7.1.0",
  "referencing==0.36.2",
  "regex==2025.11.3",
  "requests==2.32.5",
  "requests-oauthlib==2.0.0",
  "requests-toolbelt==1.0.0",
  "rich==14.2.0",
  "rich-rst==1.3.2",
  "rpds-py==0.30.0",
  "rsa==4.9.1",
  "shellingham==1.5.4",
  "six==1.17.0",
  "smithery==0.4.4",
  "sniffio==1.3.1",
  "sortedcontainers==2.4.0",
  "soupsieve==2.8",
  "SQLAlchemy==2.0.45",
  "sse-starlette==3.0.3",
  "starlette==0.50.0",
  "sympy==1.14.0",
  "tavily-python==0.7.15",
  "tenacity==8.5.0",
  "tiktoken==0.12.0",
  "tokenizers==0.22.1",
  "toml==0.10.2",
  "tqdm==4.67.1",
  "typer==0.20.0",
  "typer-slim==0.20.0",
  "typing-inspect==0.9.0",
  "typing-inspection==0.4.2",
  "typing_extensions==4.15.0",
  "uritemplate==4.2.0",
  "urllib3==2.3.0",
  "uuid_utils==0.12.0",
  "uvicorn==0.38.0",
  "uvloop==0.22.1",
  "watchfiles==1.1.1",
  "wcmatch==10.1",
  "websocket-client==1.9.0",
  "websockets==15.0.1",
  "wrapt==1.17.3",
  "xxhash==3.6.0",
  "yarl==1.22.0",
  "zipp==3.23.0",
  "zstandard==0.25.0",
]

[project.scripts]
dev = "smithery.cli.dev:main"
playground = "smithery.cli.playground:main"

[tool.smithery]
# 指向被 @smithery.server() 装饰的“工厂函数”
server = "biai_server.server:create_server"


